More cleanup, fixed hex editor stack corruption

+++ Software/GameEngine/src/games/mage/mage_defines.h
@@ -91,7 +91,7 @@ all of the old code used as the foundation of this badge.
 //these variables are reserved script and action IDs used to indicate when a script or action should not do anything.
 #define MAGE_NO_SCRIPT (-1)
 #define MAGE_NO_MAP (-1)
-#define MAGE_NO_WARP_STATE (-1)
+#define MAGE_NO_WARP_STATE ((uint16_t)-1)
 #define MAGE_NULL_SCRIPT 0
 #define MAGE_NULL_ACTION 0
